Force a screen re-draw after tcap 'ti' on terminal open
The 'tcapkopen()' function re-initializes the terminal with the 'ti'
sequence, which for most sane termcap entries is just empty.  But for
'xterm', that seems to actually be a real control sequence (clear and
reset?), and we'd better tell display.c that the screen is now garbage
and needs to be re-drawn.

Also, make tcapkclose() match the 'ti' (terminal init) with a 'te'
(terminal exit).

Maybe we should just stop playing games with ti/te, but this at least
improves the situation a bit.

Reported-by: Bijan Soleymani <bijan@psq.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>