file.c: remove crazy keyboard open/close calls
It seems to have something to do with some old DOS mode, and not having
keyboard translation on ("Insert floppy A:" questions while opening
files? Whatever).  But this is while doing normal file opens, and it is
just insane to open/close a tty across a file open.

The possible tty init/exit sequence would mess up some of the file
read/write messages.

Reported-by: Bijan Soleymani <bijan@psq.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>