Turn NBSP into regular SP on input
Particularly pasting from a web browser, I get a lot of 'space' +
'non-breaking space' noise, and keeping the &nbsp as an actual unicode
character ends up being a major pain.

Note: this is only done on input.  If the file contains the unicode
character U+00A0, we'll keep it that way.  But you can't enter it from
the keyboard (or cut-and-paste, which ends up looking like keyboard
input).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>