Allow bigger filenames
Not that I know why. 80-character filename limits are kind of cute.

Joerg suggested _POSIX_PATH_MAX, which is almost certainly the right
value.  But that's bigger than HUGE:

   #define HUGE    1000           /* Huge number                  */

which obviously means we shouldn't go quite *that* extreme.

Remember, we come from the days when it was hard to do allocations
larger than 64kB.   We have limits, dammit.

"256 bytes is enough for anybody"

Reported-by: Joerg Scheurich <mufti11@web.de>
Signed-off-by: Linus Torvalds <torvalds@linux-coundation.org>