increase size of lock table
When editing lots of files, the lock table size can become a limit.

Increase it to be HUGE.

Nobody should *ever* use this silly editor in this day and age any more.
I just can't seem to unlearn my finger memory, and it's easier for me to
do these stupid small fixes than it is to figure out how to do key
remapping in some saner editor.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>